requirements.txt
MetaTrader5
telethon
rich
httpx
python-dotenv
pandas
numpy
